chore: refactor to support 2 demos
