+++ package.json
@@ -43,6 +43,7 @@
     "class-validator": "^0.14.0",
     "cloudbees-openfeature-provider-node": "^1.0.0",
     "clsx": "^1.2.1",
+    "eventemitter3": "^5.0.1",
     "events": "^3.3.0",
     "express": "4.18.2",
     "express-validator": "^6.14.2",
@@ -53,6 +54,7 @@
     "nestjs-pino": "^3.1.2",
     "pino-http": "^8.3.1",
     "pino-pretty": "^10.0.0",
+    "react-router-dom": "^6.11.1",
     "reflect-metadata": "^0.1.13",
     "rox-browser": "^5.4.2",
     "rox-node": "^5.4.2",
@@ -62,7 +64,7 @@
     "@babel/preset-react": "^7.18.6",
     "@emotion/react": "11.11.0",
     "@emotion/styled": "11.11.0",
-    "@mui/material": "^5.11.3",
+    "@mui/material": "^5.13.1",
     "@nestjs/schematics": "9.1.0",
     "@nestjs/testing": "9.4.0",
     "@nrwl/cli": "15.9.3",
