+++ packages/ui/src/app/demos/fib3r/background.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import './background.css';
 import clsx from 'clsx';
-import { Theme } from './types';
-import { HEIGHT_MINUS_HEADER } from './constants';
+import { Theme } from '../../types';
+import { HEIGHT_MINUS_HEADER } from '../../constants';
 
 export class Background extends React.Component<{
   colors: Theme;
